refactor(login): move to user context
