+++ api/src/apps/auth/mappers/data-encode.mapper.ts
@@ -1,5 +1,5 @@
 import { UserResponse } from 'src/apps/users/dto/responses/user.response.dto';
-import { DataToEncodeDto } from '../dto/responses/data-encode.dto';
+import { DataToEncodeDto } from 'src/apps/users/dto/responses/data-encode.dto';
 
 export function mapDataToEncode(user: UserResponse): DataToEncodeDto {
   return {
